+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+* [SKB-1200] Change apply pointing to use LogicalTileId rather than extracting from the Tile's TRL.
 * [THORN-405] Re-implement intermediate healths in MccsTile, fix old health model in MccsTile.
 * [THORN-377] Use configuration properties in health.